proto: rename vtysh.proto to vty.proto
The schema has long since outgrown its original "vtysh" name — it
now drives every gRPC surface zebra-rs offers (Exec, Show, Apply,
Clear, Config, Register, ExecModule). Rename the file and the
proto package so the include_proto! / generated module name lines
up with the rest of the vty* crate family.

Mechanical sweep:
- proto/vtysh.proto → proto/vty.proto, package vtysh → package vty
- 3 build.rs files: compile_protos path
- vtyctl, vtyhelper, zebra-rs: include_proto!("vty"), mod vty,
  and every `vtysh::` use → `vty::`
