+++ vtyctl/build.rs
@@ -1,4 +1,4 @@
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    tonic_build::compile_protos("../proto/vtysh.proto")?;
+    tonic_build::compile_protos("../proto/vty.proto")?;
     Ok(())
 }
