+++ zebra-rs/build.rs
@@ -1,7 +1,7 @@
 use std::process::Command;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    tonic_build::compile_protos("../proto/vtysh.proto")?;
+    tonic_build::compile_protos("../proto/vty.proto")?;
 
     // Capture git information at build time
     set_git_info();
